refactor: rename recode_v2 module to code
- Simplified command: patina scrape code (was recode)
- Module renamed: src/commands/scrape/recode_v2/ → code/
- Removed v2 references from user-facing output
- Clean naming now that refactor is mature and battle-tested
- Tested across all repos: working perfectly
