+++ layer/sessions/20250909-090913.md
@@ -1,0 +1,63 @@
+# Session: scrape
+**ID**: 20250909-090913
+**Started**: 2025-09-09T13:09:13Z
+**LLM**: claude
+**Git Branch**: work
+**Session Tag**: session-20250909-090913-start
+**Starting Commit**: bf29d82f3405edc4369f8c3a9072b200068b3940
+
+## Previous Session Context
+The previous session successfully migrated the entire codebase from string-based CallEdge to type-safe CallGraphEntry with CallType enum, updating all 8 language processors and achieving clean commits for each. The major cleanup removed ~569 lines of dead SQL code and fixed critical CI warnings, though a philosophical decision point emerged about whether to keep unused type infrastructure (SymbolName, DocString, node constants) that aligns with identified missing features like module hierarchy and constant extraction. The session ended with the codebase ready for PR but with the unused code question unresolved.
+
+## Goals
+- [ ] Review scrape command recode v2 cleanup decisions
+- [ ] Decide whether to finish cleanup (remove unused types) or keep for future extensibility
+- [ ] Prepare final PR if cleanup approach is decided
+
+## Activity Log
+### 09:09 - Session Start
+Session initialized with goal: scrape
+Working on branch: work
+Tagged as: session-20250909-090913-start
+
+
+### 11:56 - Update (covering since 09:09)
+
+**Git Activity:**
+- Commits this session:        1
+- Files changed: 7
+- Last commit: 76 minutes ago
+
+**Work completed:**
+- Reviewed Git history and analyzed cleanup decisions from previous sessions
+- Deleted unused types (SymbolName, DocString) and CallType variants (Template, Event) - ~145 lines removed
+- Fixed clippy warnings blocking CI (empty line after doc comment)
+- Fixed flaky session tests with robust Drop guard pattern using Option<PathBuf>
+- Successfully pushed fixes to PR #24 which then passed all CI checks
+- PR #24 merged to main, completing the major refactor from SQL to ExtractedData structs
+
+**Key decisions:**
+- Chose pragmatic cleanup over preserving "future vision" infrastructure
+- Deleted unused types rather than preserving with #[allow(dead_code)]
+- Used Drop guard pattern for test directory restoration (senior Rust dev approach)
+- Made guards handle current_dir() failures gracefully with Option<PathBuf>
+
+**Challenges faced:**
+- 7 clippy errors blocking CI due to unused code
+- Flaky tests failing due to directory restoration in parallel test environment
+- Tests using global state (current_dir) causing interference
+- Fixed with Drop guards that ensure cleanup even on panic
+
+**Patterns observed:**
+- Drop guards are the idiomatic Rust way to handle test cleanup
+- Using Option<PathBuf> prevents panics when current_dir() fails
+- Pragmatic cleanup beats speculative infrastructure
+- Small focused commits make PR reviews easier
+
+
+## Session Classification
+- Work Type: pattern-work
+- Files Changed:        7
+- Commits:        1
+- Patterns Modified:        1
+- Session Tags: session-20250909-090913-start..session-20250909-090913-end
