+++ layer/sessions/20250909-122742.md
@@ -1,0 +1,71 @@
+# Session: recode v2 rename and review
+**ID**: 20250909-122742
+**Started**: 2025-09-09T16:27:42Z
+**LLM**: claude
+**Git Branch**: work
+**Session Tag**: session-20250909-122742-start
+**Starting Commit**: 4781029c5aed6d8dbaf3f704d678399b4667e5e2
+
+## Previous Session Context
+The previous session completed the major scrape command refactor by finishing pragmatic cleanup - removing ~145 lines of unused types (SymbolName, DocString, unused CallType variants) and fixing CI-blocking clippy warnings. Fixed flaky session tests using Drop guard pattern with Option<PathBuf> for robust directory restoration, then successfully merged PR #24 which migrated the entire codebase from SQL to ExtractedData structs. The session demonstrated that pragmatic cleanup beats speculative infrastructure.
+
+## Goals
+- [ ] Review the recode v2 implementation for consistency and naming
+- [ ] Consider renaming 'recode' command to better reflect its purpose
+- [ ] Ensure v2 refactor aligns with the ExtractedData patterns from scrape
+- [ ] Review command structure and identify any cleanup opportunities
+
+## Activity Log
+### 12:27 - Session Start
+Session initialized with goal: recode v2 rename and review
+Working on branch: work
+Tagged as: session-20250909-122742-start
+
+
+### 14:25 - Update (covering since 12:27)
+
+**Git Activity:**
+- Commits this session:        0
+- Files changed: 4
+- Last commit: 4 hours ago
+
+**Work completed:**
+- Reviewed complete git history and session files to understand recode v2 evolution
+- Successfully renamed recode_v2 module to code throughout entire codebase
+- Updated CLI: `patina scrape recode` → `patina scrape code` 
+- Cleaned all references: enum variants, function names, doc comments, output messages
+- Built and installed release version with cleaned naming
+- Tested across all repos: dagger (3.8s), dust (1.9s), dojo (1.4s), SDL (8s), duckdb (27s), patina (425ms)
+
+**Key decisions:**
+- Returned to original "code" name now that refactor is mature and battle-tested
+- "recode" was transitional marking the rewrite phase - no longer needed
+- Kept it simple: "code" is clearer than "recode" for semantic extraction
+- Removed all "v2" references from user-facing output (was showing implementation detail)
+
+**Challenges faced:**
+- Git showed massive deletions (6147 lines) but these were just renames of recode_v2/ → code/
+- Found lingering references in doc comments and output messages
+- All cleaned up for professional presentation
+
+**Patterns observed:**
+- The rename journey: code.rs (monolith) → recode_v2 (clean rewrite) → code (mature)
+- Naming should reflect purpose not history - users don't need to know about v2
+- Simple names are better when the implementation is solid
+- All test repos work perfectly with renamed command
+
+
+### 14:26 - Update (covering since 14:25)
+
+**Git Activity:**
+- Commits this session:        0
+- Files changed: 4
+- Last commit: 4 hours ago
+
+
+## Session Classification
+- Work Type: exploration
+- Files Changed:        0
+- Commits:        0
+- Patterns Modified:        0
+- Session Tags: session-20250909-122742-start..session-20250909-122742-end
