+++ src/commands/scrape/code/languages/typescript.rs
@@ -17,11 +17,11 @@
 //! - Decorators and metadata
 //! - JSX support (.tsx files with different parser)
 
-use crate::commands::scrape::recode_v2::database::{
+use crate::commands::scrape::code::database::{
     CodeSymbol, FunctionFact, ImportFact, TypeFact,
 };
-use crate::commands::scrape::recode_v2::extracted_data::ExtractedData;
-use crate::commands::scrape::recode_v2::types::{CallGraphEntry, CallType, FilePath};
+use crate::commands::scrape::code::extracted_data::ExtractedData;
+use crate::commands::scrape::code::types::{CallGraphEntry, CallType, FilePath};
 use anyhow::{Context, Result};
 use tree_sitter::{Node, Parser};
