+++ src/main.rs
@@ -114,7 +114,7 @@ struct ScrapeArgs {
 #[derive(Subcommand)]
 enum ScrapeCommands {
     /// Extract semantic information using modular architecture
-    Recode {
+    Code {
         #[command(flatten)]
         args: ScrapeArgs,
     },
@@ -270,7 +270,7 @@ fn main() -> Result<()> {
         },
         Commands::Scrape { command } => {
             // Default to Code subcommand with default args for backward compatibility
-            let subcommand = command.unwrap_or(ScrapeCommands::Recode {
+            let subcommand = command.unwrap_or(ScrapeCommands::Code {
                 args: ScrapeArgs {
                     init: false,
                     query: None,
@@ -279,8 +279,8 @@ fn main() -> Result<()> {
                 },
             });
             match subcommand {
-                ScrapeCommands::Recode { args } => {
-                    commands::scrape::execute_recode(args.init, args.query, args.repo, args.force)?;
+                ScrapeCommands::Code { args } => {
+                    commands::scrape::execute_code(args.init, args.query, args.repo, args.force)?;
                 }
                 ScrapeCommands::Docs { args } => {
                     commands::scrape::execute_docs(args.init, args.query, args.repo, args.force)?;
